requirements.txt
Django==5.1
mysqlclient
djangorestframework==3.15.2
django-cors-headers
django-environ
